+++ src/main/kotlin/slang/plugin/lsp/lsp4ij/SlangLanguageServer.kt
@@ -10,6 +10,6 @@ class SlangLanguageServer(project: Project): OSProcessStreamConnectionProvider()
     init {
         val provider = SlangLanguageServerProvider.getInstance(project)
         assert(provider.checkValidLanguageServerFiles())
-        commandLine = GeneralCommandLine(provider.getSlangDExecutablePath().absolutePathString(), "--stdio")
+        commandLine = GeneralCommandLine(provider.getSlangDExecutablePath().absolutePathString(), "--stdio", "--debug")
     }
 }